# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AIS API",
        "description": "AIS API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/product/{id}/edit": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Retrieve a product for editing",
                "operationId": "0dd7e5c08d75ce8c3c93b1c5f7241472",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nameTh": {
                                                    "type": "string",
                                                    "example": "ตู้เย็น LG"
                                                },
                                                "nameEn": {
                                                    "type": "string",
                                                    "example": "LG Fridge"
                                                },
                                                "serialNumber": {
                                                    "type": "string",
                                                    "example": "INVFR0000001"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "double",
                                                    "example": 10000
                                                },
                                                "category": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "nameTh": {
                                                            "type": "string",
                                                            "example": "ตู้เย็น"
                                                        },
                                                        "nameEn": {
                                                            "type": "string",
                                                            "example": "Fridge"
                                                        },
                                                        "prefixSerialNumber": {
                                                            "type": "string",
                                                            "example": "FR"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "400": {
                        "description": "Validation input error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}